requirements.txt
impacket>=0.10.0
ldap3>=2.9
psycopg2-binary>=2.9.5
python-dotenv>=0.19.0
rich>=13.7.0
typer>=0.9.0
colorama>=0.4.6
shellingham>=1.5.4
typing-extensions>=4.8.0
markdown-it-py>=3.0.0
mdurl>=0.1.2
pygments>=2.17.2 